Move SamplingStatus to spi package
