+++ jaeger-core/src/main/java/io/jaegertracing/internal/samplers/GuaranteedThroughputSampler.java
@@ -16,6 +16,7 @@ package io.jaegertracing.internal.samplers;
 
 import io.jaegertracing.internal.Constants;
 import io.jaegertracing.spi.Sampler;
+import io.jaegertracing.spi.SamplingStatus;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.EqualsAndHashCode;
