+++ jaeger-core/src/main/java/io/jaegertracing/internal/samplers/RateLimitingSampler.java
@@ -17,6 +17,7 @@ package io.jaegertracing.internal.samplers;
 import io.jaegertracing.internal.Constants;
 import io.jaegertracing.internal.utils.RateLimiter;
 import io.jaegertracing.spi.Sampler;
+import io.jaegertracing.spi.SamplingStatus;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
