+++ jaeger-core/src/main/java/io/jaegertracing/internal/samplers/RemoteControlledSampler.java
@@ -23,6 +23,7 @@ import io.jaegertracing.internal.samplers.http.RateLimitingSamplingStrategy;
 import io.jaegertracing.internal.samplers.http.SamplingStrategyResponse;
 import io.jaegertracing.spi.Sampler;
 import io.jaegertracing.spi.SamplingManager;
+import io.jaegertracing.spi.SamplingStatus;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
