+++ jaeger-core/src/test/java/io/jaegertracing/internal/samplers/ProbabilisticSamplerTest.java
@@ -19,6 +19,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import io.jaegertracing.spi.Sampler;
+import io.jaegertracing.spi.SamplingStatus;
 import java.util.Map;
 import org.junit.Test;
