+++ jaeger-core/src/test/java/io/jaegertracing/internal/samplers/TestConstSampler.java
@@ -16,6 +16,7 @@ package io.jaegertracing.internal.samplers;
 
 import static org.junit.Assert.assertEquals;
 
+import io.jaegertracing.spi.SamplingStatus;
 import java.util.Map;
 import org.junit.Test;
